pyproject.toml
[project]
name = "supabase-schema-mcp"
version = "0.1.0"
description = "MCP server for deep Supabase/Postgres schema introspection (tables, RLS, functions, relationships, triggers)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "asyncpg",
    "httpx",
    "python-dotenv",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
supabase-schema-mcp = "supabase_schema_mcp.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/supabase_schema_mcp"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
